+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/AutoParking.java
@@ -8,8 +8,6 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
-import static org.firstinspires.ftc.teamcode.autonomous.ParkingLocationAnalyzer.ParkingLocation;
-
 @Config
 @Autonomous(name="Autonomous Parking", group="Autonomous")
 public class AutoParking extends LinearOpMode {
