+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/ParkingLocation.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+public enum ParkingLocation {
+    LEFT,
+    CENTER,
+    RIGHT
+}
