+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/GamepadButton.java
@@ -1,0 +1,25 @@
+package org.firstinspires.ftc.teamcode.opmode;
+
+public enum GamepadButton {
+    A,
+    B,
+    X,
+    Y,
+    DPAD_UP,
+    DPAD_DOWN,
+    DPAD_LEFT,
+    DPAD_RIGHT,
+    LEFT_BUMPER,
+    RIGHT_BUMPER,
+    LEFT_STICK_BUTTON,
+    RIGHT_STICK_BUTTON,
+    GUIDE,
+    START,
+    BACK,
+    LEFT_TRIGGER,
+    RIGHT_TRIGGER,
+    LEFT_STICK_X,
+    LEFT_STICK_Y,
+    RIGHT_STICK_X,
+    RIGHT_STICK_Y
+}
