+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/MeetTwo.java
@@ -21,9 +21,6 @@ import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigu
 import java.util.Arrays;
 import java.util.List;
 
-import static org.firstinspires.ftc.teamcode.opmode.GamepadInterface.GamepadController;
-import static org.firstinspires.ftc.teamcode.opmode.GamepadInterface.GamepadButton;
-
 @Config
 @TeleOp(name="Meet Two", group="TeleOp")
 public class MeetTwo extends LinearOpMode {
